refactor(cognition): move pure perception-predicate tests into perception.rs (pure code-motion)
Relocated the three self-contained perception tests — mutation_is_read... (mutated_workspace),
unacted_claim_geometry (claimed_file_without_act), unobserved_mutation_geometry
(wrote_without_observation) — into perception.rs's own `#[cfg(test)] mod tests`. Left in mod.rs
the two whose attribution is not clean: the_orientation_counter (exercises WorkingMemory, not a
predicate) and redundant_orientation_fires (depends on the shared tool_call() harness helper).
Trimmed the now-unused perception imports from mod.rs's test imports. Behavior-identical.
mod.rs 1355 → 1255 lines; perception.rs 257 → 365 lines. act_observe tests green: 28 passed, 0 failed.
